+++ CHANGELOG.md
@@ -3,20 +3,29 @@ CHANGELOG
 
 0.44.0
 ------
-- (Experimental) Sixel image support in preview window
+- (Experimental) Sixel image support in preview window (not available on Windows)
     - [bin/fzf-preview.sh](bin/fzf-preview.sh) is added to demonstrate how to
       display an image using Kitty image protocol or Sixel. You can use it
       like so:
       ```sh
       fzf --preview='fzf-preview.sh {}'
       ```
-- (Experimental) Sixel and Kitty image support now also available on Windows
+- (Experimental) iTerm2 inline image protocol support in preview window (not available on Windows)
+  ```sh
+  # Using https://iterm2.com/utilities/imgcat
+  fzf --preview 'imgcat -W $FZF_PREVIEW_COLUMNS -H $FZF_PREVIEW_LINES {}'
+  ```
 - HTTP server can be configured to accept remote connections
   ```sh
   # FZF_API_KEY is required for a non-localhost listen address
   export FZF_API_KEY="$(head -c 32 /dev/urandom | base64)"
   fzf --listen 0.0.0.0:6266
   ```
+    - To allow remote process execution, use `--listen-unsafe` instead
+      (`execute*`, `reload*`, `become`, `preview`, `change-preview`, `transform-*`)
+      ```sh
+      fzf --listen-unsafe 0.0.0.0:6266
+      ```
 - Bug fixes
 
 0.43.0